feat: refactoring folder AlbumCard
