+++ src/components/common/AlbumCard.jsx
@@ -5,7 +5,7 @@ import { Link } from "react-router-dom";
 const AlbumCard = ({ album }) => {
   return (
     <Link to={`/album/${album.id}`} style={{ textDecoration: "none" }}>
-      <Card sx={{ overflow: "hidden", border: 'none', boxShadow: 'none' }}> 
+      <Card sx={{ overflow: "hidden", border: 'none', boxShadow: 'none', padding: '10px' }}> 
         <CardMedia
           component="img"
           image={album.cover}
